Allow partial wrapper initialisation for early access
