+++ AndroidSDKCore/src/main/java/com/leanplum/migration/wrapper/NullWrapper.kt
@@ -1,6 +1,6 @@
 package com.leanplum.migration.wrapper
 
 /**
- * Purpose of this class is to be used instead of null value to avoid checking for null in Java.
+ * Purpose of this singleton is to be used instead of null value to avoid checking for null in Java.
  */
 internal object NullWrapper : IWrapper
